Prevent System.ext from having a 'constructor'
The System.ext configuration is used to support Steal's ability to load
plugins without the bang. Since it was previously a plain object, it had
a constructor property. This meant if you tried to do:
`require("./foo.constructor");` where this is actually
`foo.constructor.js`, it would think that it needed to use `ext`
configuration.

The fix is to use `Object.create(null)` which creates an object with no
properties. Closes #938